feat: changed a lot of things.

+++ src/domain.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 pub struct TokenResponse {
     pub access_token: String,
     // token_type: String,
-    pub expires_in: u32,
+    pub expires_in: u64,
     pub refresh_token: String,
     // scope: String,
 }
